refactor: final cleanup - move sidebars and remove empty folders
Completed component restructuring with final cleanup:

1. Moved domain-specific sidebars to feature folders:
   - sidebar/ProjectSidebar.tsx → feature/project/ProjectSidebar.tsx
   - sidebar/TeamSidebar.tsx → feature/team/TeamSidebar.tsx

2. Removed empty and obsolete folders:
   - dropdown/ (all files moved to ui/dropdown/)
   - system/ (text.css moved to ui/text/)
   - sidebar/ (all files moved to feature/)

3. Updated import paths:
   - ProjectPage.tsx: Updated ProjectSidebar import
   - Team.tsx: Updated TeamSidebar import
   - globals.css: Updated text.css import path

Final component structure:
- ui/ - Reusable UI components
- card/ - Card components by domain
- layout/ - Layout components by function
- section/ - Section components by domain
- feature/ - Feature-specific components by domain
- modal/ - Modal components and context
- toast/ - Toast components and context

All changes verified with pnpm lint and pnpm build.
