+++ src/app/(box-layout)/team/Team.tsx
@@ -2,7 +2,7 @@
 import { getTeamData } from '@/services/team/getTeamData.server';
 import { getTeamProjects } from '@/services/team/getTeamProjects.server';
 import { notFound } from 'next/navigation';
-import TeamSidebar from '@/app/components/sidebar/TeamSidebar';
+import TeamSidebar from '@/app/components/feature/team/TeamSidebar';
 import ProfileIcon from '@/app/components/card/portfolio/components/ProfileIcon';
 import ProjectGrid from '@/app/components/feature/project/components/ProjectGrid';
 import SidebarContentLayout from '@/app/components/layout/sidebar/SidebarContentLayout';
