--- src/app/components/system/text.css
@@ -1,62 +1,0 @@
-.text-title-en {
-  @apply font-bold; /* uses custom font-threat via global font-face / utility */
-  font-size: 1.5rem; /* 24px */
-  line-height: 1.5rem; /* 24px */
-  letter-spacing: -0.023em;
-}
-
-.text-title {
-  @apply font-bold;
-  font-size: 1.5rem; /* 24px */
-  line-height: 2rem; /* 32px */
-  letter-spacing: -0.023em;
-}
-
-.text-heading2 {
-  @apply font-semibold;
-  font-size: 1.25rem; /* 20px */
-  line-height: 2.25rem; /* 36px */
-  letter-spacing: -0.012em;
-}
-
-.text-heading {
-  @apply font-semibold;
-  font-size: 1.25rem; /* 20px */
-  line-height: 1.5rem; /* 24px */
-  letter-spacing: -0.012em;
-}
-
-.text-body {
-  @apply font-normal;
-  font-size: 1rem; /* 16px */
-  line-height: 1.375rem; /* 22px */
-  letter-spacing: 0.0057em;
-}
-
-.text-body2 {
-  @apply font-semibold;
-  font-size: 1rem; /* 16px */
-  line-height: 1.375rem; /* 22px */
-  letter-spacing: 0.0057em;
-}
-
-.text-label2 {
-  @apply font-medium;
-  font-size: 0.875rem; /* 14px */
-  line-height: 1.25rem; /* 20px */
-  letter-spacing: 0.0145em;
-}
-
-.text-label {
-  @apply font-medium;
-  font-size: 0.875rem; /* 14px */
-  line-height: 1.25rem; /* 20px */
-  letter-spacing: 0.0145em;
-}
-
-.text-caption {
-  @apply font-normal;
-  font-size: 0.75rem; /* 12px */
-  line-height: 1rem; /* 16px */
-  letter-spacing: 0.0252em;
-}
